refactor(uniquely/salavat-pwa): structure
