+++ uniquely/salavat-pwa/index.html
@@ -50,7 +50,7 @@
   <body>
     <alwatr-pwa-root></alwatr-pwa-root>
 
-    <script type="module" src="build/pwa-root.js"></script>
+    <script type="module" src="build/components/pwa-root.js"></script>
 
     <noscript>
       <div class="noscript">
